# backup: add simpler RESTORE span computation