--- a/pkg/ccl/backupccl/bench_covering_test.go
+++ b/pkg/ccl/backupccl/bench_covering_test.go
@@ -12,14 +12,13 @@ import (
 	"context"
 	fmt "fmt"
 	"testing"
-	"time"
 
-	"github.com/cockroachdb/cockroach/pkg/roachpb"
-	"github.com/cockroachdb/cockroach/pkg/util/encoding"
-	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/randutil"
 )
 
 func BenchmarkCoverageChecks(b *testing.B) {
+	r, _ := randutil.NewTestRand()
+
 	for _, numBackups := range []int{1, 7, 24, 24 * 4} {
 		b.Run(fmt.Sprintf("numBackups=%d", numBackups), func(b *testing.B) {
 			for _, numSpans := range []int{10, 20, 100} {
@@ -27,45 +26,10 @@ func BenchmarkCoverageChecks(b *testing.B) {
 					for _, baseFiles := range []int{0, 10, 100, 1000, 10000} {
 						b.Run(fmt.Sprintf("numFiles=%d", baseFiles), func(b *testing.B) {
 							b.StopTimer()
-							backups := make([]BackupManifest, numBackups)
-							ts := hlc.Timestamp{WallTime: time.Second.Nanoseconds()}
-							for i := range backups {
-								backups[i].Spans = make(roachpb.Spans, numSpans)
-								for j := range backups[i].Spans {
-									backups[i].Spans[j] = makeTableSpan(uint32(100 + j + (i / 4)))
-								}
-								backups[i].EndTime = ts.Add(time.Minute.Nanoseconds()*int64(i), 0)
-								if i > 0 {
-									backups[i].StartTime = backups[i-1].EndTime
-									if i%4 == 0 {
-										backups[i].IntroducedSpans = roachpb.Spans{backups[i].Spans[numSpans-1]}
-									}
-								}
-
-								numFiles := baseFiles
-								if i == 0 {
-									backups[i].Files = make([]BackupManifest_File, numFiles)
-								} else {
-									numFiles = numFiles / 2
-									backups[i].Files = make([]BackupManifest_File, numFiles)
-								}
-
-								for f := range backups[i].Files {
-									backups[i].Files[f].Path = fmt.Sprintf("1234567890%d.sst", f)
-									backups[i].Files[f].Span.Key = encoding.EncodeVarintAscending(backups[i].Spans[f*numSpans/numFiles].Key, int64(f))
-									backups[i].Files[f].Span.EndKey = encoding.EncodeVarintAscending(backups[i].Spans[f*numSpans/numFiles].Key, int64(f+1))
-
-								}
-								backups[i].Dir = roachpb.ExternalStorage{S3Config: &roachpb.ExternalStorage_S3{
-									Bucket:    "some-string-name",
-									Prefix:    "some-string-path/to/some/file",
-									AccessKey: "some-access-key",
-									Secret:    "some-secret-key",
-								}}
-							}
+							ctx := context.Background()
+							backups := MockBackupChain(numBackups, numSpans, baseFiles, r)
 							b.ResetTimer()
 
-							ctx := context.Background()
 							b.Run("checkCoverage", func(b *testing.B) {
 								b.ResetTimer()
 								for i := 0; i < b.N; i++ {