--- a/pkg/ccl/backupccl/bench_covering_test.go
+++ b/pkg/ccl/backupccl/bench_covering_test.go
@@ -57,3 +57,42 @@ func BenchmarkCoverageChecks(b *testing.B) {
 		})
 	}
 }
+
+func BenchmarkRestoreEntryCover(b *testing.B) {
+	r, _ := randutil.NewTestRand()
+
+	for _, numBackups := range []int{1, 2, 24, 24 * 4} {
+		b.Run(fmt.Sprintf("numBackups=%d", numBackups), func(b *testing.B) {
+			for _, baseFiles := range []int{0, 100, 10000} {
+				b.Run(fmt.Sprintf("numFiles=%d", baseFiles), func(b *testing.B) {
+					for _, numSpans := range []int{10, 100} {
+						b.Run(fmt.Sprintf("numSpans=%d", numSpans), func(b *testing.B) {
+							b.StopTimer()
+							ctx := context.Background()
+							backups := MockBackupChain(numBackups, numSpans, baseFiles, r)
+							b.ResetTimer()
+							b.Run("simple", func(b *testing.B) {
+								for i := 0; i < b.N; i++ {
+									if err := checkCoverage(ctx, backups[numBackups-1].Spans, backups); err != nil {
+										b.Fatal(err)
+									}
+									cov := makeSimpleImportSpans(backups[numBackups-1].Spans, backups, nil, nil)
+									b.ReportMetric(float64(len(cov)), "coverSize")
+								}
+							})
+							b.Run("coveringMerge", func(b *testing.B) {
+								for i := 0; i < b.N; i++ {
+									cov, _, err := makeImportSpans(backups[numBackups-1].Spans, backups, nil, nil, errOnMissingRange)
+									if err != nil {
+										b.Fatal(err)
+									}
+									b.ReportMetric(float64(len(cov)), "coverSize")
+								}
+							})
+						})
+					}
+				})
+			}
+		})
+	}
+}