Previously we used `overlappingCoveringMerge` during `RESTORE` to do two things:
a) compute a covering that partitioned the restoring spans and assigned files to be restored to those
partitions, and
b) verify that the key spans being restored were covered over all time spans by the backups.

The former produced a much more granularly partitioned covering than was
required, splitting the overall covering at all file boundaries, in all layers.
When the spans of files in incremental layers did not line up with the spans in other 
layers, this produced a huge number of tiny partitions, with each file having to
be copied to many, many partitions it overlapped.

Instead, this change introduces a new method of deriving this covering. It now
uses the files of the first backup that cover any part of the key span being restored
to define partitions and then assigns later files to any partitions they overlap. This 
produces only one partition per file that increases the covered spans, so if all the
inc backups are of the same tables/spans, it'd just be file spans of the first backup.

This simpler approach no longer has the side effect of also validating that all the
key spans are covered at all time spans that the overlappingCoveringMerge did,
so when used the caller needs to also perform their own coverage verification check,
which can use the cheaper method previously introduced in #74393.


There are more extensive benchmark results the last commit in this PR, but for a quick summary:
```
name                                                         old time/op    new time/op    delta
RestoreEntryCover/numBackups=1/numFiles=10000/numSpans=100     48.5ms ± 2%    26.3ms ± 1%   -45.68%  (p=0.008 n=5+5)
RestoreEntryCover/numBackups=24/numFiles=10000/numSpans=100     1.26s ± 1%     0.33s ± 1%   -74.21%  (p=0.008 n=5+5)
RestoreEntryCover/numBackups=96/numFiles=10000/numSpans=100     9.80s ± 1%     1.47s ± 1%   -84.99%  (p=0.036 n=3+5)

name                                                         old alloc/op   new alloc/op   delta
RestoreEntryCover/numBackups=1/numFiles=10000/numSpans=100     44.9MB ± 0%     8.9MB ± 0%   -80.21%  (p=0.008 n=5+5)
RestoreEntryCover/numBackups=24/numFiles=10000/numSpans=100     917MB ± 0%      43MB ± 0%   -95.31%  (p=0.008 n=5+5)
RestoreEntryCover/numBackups=96/numFiles=10000/numSpans=100    4.93GB ± 0%    0.24GB ± 0%   -95.06%  (p=0.000 n=3+4)
```

Release note (performance improvement): The merging of incremental backup layers during RESTORE now uses a simpler and less memory intensive algorithm.