@dt 2022-01-03
First commit is #74394. Need to write some tests and then benchmark it too but wanted to get this out for initial look.